JavaFX & MVC Improvements
The structure is being improved to fit the concept of MVC. Although it
is not yet fully flexible, it is getting better.

The GUI is fully working with current basic model/business logic;
retrieving and uploading data to/from Firebase DB.

The very important thing to consider is that the fx:controller for
JavaFX view, is now not assigned to FXML file by default, but rather
appended to it runtime.