Scene Management
The commit provides a set of tools that will make
the scene management easier.

JavaFxApplication now contains a logic for initialization
of all scenes that will be available in the application.
It is needed for more centralized place where changes can be made.
It avoids the mess across the classes.

JavaFXView is introduced as a base class for views. At the moment,
it only contains a switchScene method that is used by majority
of the views.

Minor changes and refactors to fxml, css and controller files.