Connecting GUI and Feature logic
The commit proposes a set of implementations to connect the GUI
part of the application together with the corresponding logic.

The Veg Meal controller, view and interface is also introduced for
the purpose.

The application can now handle authentication with redirecting to
the homescreen if the credentials are valid. It also
loads basic user information for the labels in the homescreen view.