Refactored UserProfile
A separate class for singleton access to UserProfile was removed.
Now singleton pattern is established inside UserProfile.

The data points are excluded from UserProfile to a separate class
UserData. This is done for server to have access to the same object,
so it can communicate with client more precisely.

Tests, controllers and other models are modified according to refactored
UserProfile.