Server: Firebase Tests & Bug Fix
Some tests are added to test connection & logic with Firebase services.

The CountDownLatch, that is used for syncing the database connection,
is supposed to have a count down method after the completion of
the desired methods. Previously, it was called before causing
Null exception.

Checkstyle violations for Firebase classes are also resolved here.