Bug Fixes: Logout, Sign In Issue handling
The logout button in the homescreen was not working. The commit solves
the issue.

When the server was disabled, the sign in procedure would lead to
home screen with empty fields. The commit produces a fix that
prevents user of getting to home screen if a proper response
was not received.

Window title is changed to something more general.