Remove secret keys from the project
The secret keys that are required for the CoolClimate API and
Firebase services are removed from the project. Although they
can be seen in the git history, they will be globally deactivated for
security purposes.

Moreover, the keys constants are refactored; moved into a single config
file at ServerApi.java, where they can be configured easily.